Add command,lookup table and parser
Added commands with child classes to abstract away multiple classes so
when teammates implement their own classes is easier. Same thing with
Parser. For LookUptable, I realised a generic execute command can take
in a hashmap and then depending on which class, or type of parser it
will use whichever list it needs to modify respectively.